Move over resample-data to new CLI group
